tests: rename provider mutation tests
